+++ test/TestBuybackFOT.t.sol
@@ -304,6 +304,7 @@ contract TestBuybackFOT is Test {
                 controller,
                 uint256(0),
                 1e18,
+                payValue,
                 int24(0),
                 uint128(0),
                 bytes32(0),