+++ test/TestOracleRevertBehavior.t.sol
@@ -341,6 +341,7 @@ contract TestOracleRevertBehavior is Test {
                 controller,
                 uint256(0),
                 1e18,
+                payAmount,
                 int24(0),
                 uint128(0),
                 bytes32(0),