+++ test/regression/BuySideFOTProjectTokenDoS.t.sol
@@ -269,6 +269,7 @@ contract RegressionBuySideFOTProjectTokenDoSTest is Test {
                 IJBController(address(controller)),
                 uint256(0),
                 1e18,
+                amountIn,
                 int24(0),
                 uint128(0),
                 bytes32(0),